Refactor Vector code to remove use of vec3_t Remove duplicate Vector definitions in util_vector.h Remove shared_vec3_origin Remove unused vector functions

+++ dlls/vector.h
@@ -15,6 +15,8 @@
 #ifndef VECTOR_H
 #define VECTOR_H
 
+#include <math.h>
+
 //=========================================================
 // 2DVector - used for many pathfinding and many other 
 // operations that are treated as planar rather than 3d.
@@ -29,7 +31,7 @@ public:
 	inline Vector2D operator*(float fl)				const	{ return Vector2D(x*fl, y*fl);	}
 	inline Vector2D operator/(float fl)				const	{ return Vector2D(x/fl, y/fl);	}
 	
-	inline float Length()						const	{ return sqrt(x*x + y*y );		}
+	inline float Length()						const	{ return static_cast<float>(sqrt(x*x + y*y )); }
 
 	inline Vector2D Normalize () const
 	{
@@ -78,7 +80,7 @@ public:
 	
 	// Methods
 	inline void CopyToArray(float* rgfl) const		{ rgfl[0] = x, rgfl[1] = y, rgfl[2] = z; }
-	inline float Length() const					{ return sqrt(x*x + y*y + z*z); }
+	inline float Length() const					{ return static_cast<float>(sqrt(x*x + y*y + z*z)); }
 	operator float *()								{ return &x; } // Vectors will now automatically convert to float * when needed
 	operator const float *() const					{ return &x; } // Vectors will now automatically convert to float * when needed
 	inline Vector Normalize() const
@@ -98,7 +100,7 @@ public:
 
 		return Vec2;
 	}
-	inline float Length2D() const					{ return sqrt(x*x + y*y); }
+	inline float Length2D() const					{ return static_cast<float>(sqrt(x*x + y*y)); }
 
 	// Members
 	vec_t x, y, z;
